requirements.txt
# YouTube Shorts Trend Analysis MVP Dependencies

# Core AI framework
pydantic-ai[anthropic,openai,gemini]

# HTTP client for API calls
httpx

# Configuration and environment management
python-dotenv
pydantic-settings

# Rate limiting for API requests
aiometer

# Testing framework
pytest
pytest-asyncio
pytest-cov

# Type checking
mypy

# Code formatting and linting
ruff

# Additional utilities
typing-extensions

# System monitoring and operations
psutil

# Process management
supervisor

# Enhanced HTTP client
uvicorn[standard]

# Future web interface framework
fastapi

# Redis client for caching
redis

# Google Generative AI for video analysis
google-generativeai